Remove overlined test scenes
