--- osu.Game.Tests/Visual/Multiplayer/TestSceneOverlinedParticipants.cs
@@ -1,57 +1,0 @@
-// Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
-// See the LICENCE file in the repository root for full licence text.
-
-using NUnit.Framework;
-using osu.Framework.Graphics;
-using osu.Game.Screens.Multi.Components;
-using osu.Game.Users;
-
-namespace osu.Game.Tests.Visual.Multiplayer
-{
-    public class TestSceneOverlinedParticipants : MultiplayerTestScene
-    {
-        [SetUp]
-        public new void Setup() => Schedule(() =>
-        {
-            Room.RoomID.Value = 7;
-
-            for (int i = 0; i < 50; i++)
-            {
-                Room.RecentParticipants.Add(new User
-                {
-                    Username = "peppy",
-                    Id = 2
-                });
-            }
-        });
-
-        [Test]
-        public void TestHorizontalLayout()
-        {
-            AddStep("create component", () =>
-            {
-                Child = new ParticipantsDisplay(Direction.Horizontal)
-                {
-                    Anchor = Anchor.Centre,
-                    Origin = Anchor.Centre,
-                    Width = 0.2f,
-                };
-            });
-        }
-
-        [Test]
-        public void TestVerticalLayout()
-        {
-            AddStep("create component", () =>
-            {
-                Child = new ParticipantsDisplay(Direction.Vertical)
-                {
-                    Anchor = Anchor.Centre,
-                    Origin = Anchor.Centre,
-                    Width = 0.2f,
-                    Height = 0.2f,
-                };
-            });
-        }
-    }
-}
