+++ osu.Game.Tests/Visual/Multiplayer/TestSceneParticipantsList.cs
@@ -4,22 +4,54 @@
 using NUnit.Framework;
 using osu.Framework.Graphics;
 using osu.Game.Screens.Multi.Components;
+using osu.Game.Users;
 
 namespace osu.Game.Tests.Visual.Multiplayer
 {
     public class TestSceneParticipantsList : MultiplayerTestScene
     {
-        protected override bool UseOnlineAPI => true;
-
-        public TestSceneParticipantsList()
-        {
-            Add(new ParticipantsList { RelativeSizeAxes = Axes.Both });
-        }
-
         [SetUp]
         public new void Setup() => Schedule(() =>
         {
             Room.RoomID.Value = 7;
+
+            for (int i = 0; i < 50; i++)
+            {
+                Room.RecentParticipants.Add(new User
+                {
+                    Username = "peppy",
+                    Id = 2
+                });
+            }
         });
+
+        [Test]
+        public void TestHorizontalLayout()
+        {
+            AddStep("create component", () =>
+            {
+                Child = new ParticipantsDisplay(Direction.Horizontal)
+                {
+                    Anchor = Anchor.Centre,
+                    Origin = Anchor.Centre,
+                    Width = 0.2f,
+                };
+            });
+        }
+
+        [Test]
+        public void TestVerticalLayout()
+        {
+            AddStep("create component", () =>
+            {
+                Child = new ParticipantsDisplay(Direction.Vertical)
+                {
+                    Anchor = Anchor.Centre,
+                    Origin = Anchor.Centre,
+                    Width = 0.2f,
+                    Height = 0.2f,
+                };
+            });
+        }
     }
 }
